requirements.txt
streamlit
pandas
google-generativeai
spacy
ja-ginza
wordcloud
networkx
matplotlib